Moved paginator utils
