+++ src/main/java/com/honiism/discord/lemi/utils/buttons/Paginator.java
@@ -17,7 +17,7 @@
  * along with Lemi-Bot. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.honiism.discord.lemi.utils.paginator;
+package com.honiism.discord.lemi.utils.buttons;
 
 import java.util.Collections;
 import java.util.HashSet;
@@ -25,8 +25,12 @@ import java.util.List;
 import java.util.Set;
 import java.util.concurrent.TimeUnit;
 
+import java.awt.*;
+import java.time.Instant;
+
 import javax.annotation.Nonnull;
 
+import com.honiism.discord.lemi.utils.misc.EmbedUtils;
 import com.honiism.discord.lemi.utils.misc.Emojis;
 import com.jagrosh.jdautilities.commons.waiter.EventWaiter;
 
@@ -43,36 +47,54 @@ import net.dv8tion.jda.api.requests.restaction.MessageAction;
 import net.dv8tion.jda.api.requests.restaction.WebhookMessageAction;
 import net.dv8tion.jda.internal.utils.Checks;
 
-public class EmbedPaginator {
-
+public class Paginator {
+        
+    private static Button first = Button.secondary("first", Emoji.fromMarkdown(Emojis.SKIP_TO_START_BUTTON));
+    private static Button previous = Button.secondary("previous", Emoji.fromMarkdown(Emojis.LEFT_ARROW));
+    private static Button next = Button.secondary("next", Emoji.fromMarkdown(Emojis.RIGHT_ARROW));
+    private static Button last = Button.secondary("last", Emoji.fromMarkdown(Emojis.SKIP_TO_END_BUTTON));
+    private static Button delete = Button.danger("stop", Emoji.fromMarkdown(Emojis.CROSS_MARK));
+    
     private final EventWaiter waiter;
+    private final int itemsPerPage;
     private final int pages;
     private final long timeout;
-    private final List<EmbedBuilder> items;
+    private final List<String> items;
     private final JDA jda;
     private final Set<Long> allowedUsers;
+    private final boolean numbered;
     private final String msgContent;
+    private final Color color;
     private final String footer;
-
-    private Button first = Button.secondary("first", Emoji.fromMarkdown(Emojis.SKIP_TO_START_BUTTON));
-    private Button previous = Button.secondary("previous", Emoji.fromMarkdown(Emojis.LEFT_ARROW));
-    private Button next = Button.secondary("next", Emoji.fromMarkdown(Emojis.RIGHT_ARROW));
-    private Button last = Button.secondary("last", Emoji.fromMarkdown(Emojis.SKIP_TO_END_BUTTON));
-    private Button delete = Button.danger("stop", Emoji.fromMarkdown(Emojis.CROSS_MARK));
+    private final String embedTitle;
+    private final String thumbnailUrl;
+    private final String embedDesc;
+    private final String imageUrl;
+    private final boolean timestamp;
 
     private int page = 1;
     private boolean interactionStopped = false;
 
-    private EmbedPaginator(EventWaiter waiter, long timeout, List<EmbedBuilder> items, JDA jda,
-                           Set<Long> allowedUsers, String msgContent, String footer) {
+    private Paginator(EventWaiter waiter, long timeout, List<String> items, JDA jda,
+                      Set<Long> allowedUsers, int itemsPerPage, boolean numberedItems, String msgContent,
+                      Color color, String footer, String embedTitle, String thumbnailUrl, String embedDesc,
+                      String imageUrl, boolean timestamp) {
         this.waiter = waiter;
         this.timeout = timeout;
         this.items = items;
         this.jda = jda;
         this.allowedUsers = Collections.unmodifiableSet(allowedUsers);
+        this.itemsPerPage = itemsPerPage;
+        this.numbered = numberedItems;
         this.msgContent = msgContent;
+        this.color = color;
         this.footer = footer;
-        this.pages = items.size();
+        this.pages = (int) Math.ceil((double) items.size() / itemsPerPage);
+        this.embedTitle = embedTitle;
+        this.thumbnailUrl = thumbnailUrl;
+        this.embedDesc = embedDesc;
+        this.imageUrl = imageUrl;
+        this.timestamp = timestamp;
     }
 
     public void paginate(Message message, int page) {
@@ -138,7 +160,7 @@ public class EmbedPaginator {
                     if (messageId != event.getMessageIdLong()) {
                         return false;
                     }
-
+                    
                     if (allowedUsers.size() >= 1) {
                         if (!allowedUsers.contains(event.getUser().getIdLong())) {
                             event.deferEdit().queue(s -> {}, e -> {});
@@ -156,7 +178,7 @@ public class EmbedPaginator {
                             if (page < 1) {
                                 page = 1;
                             }
-
+                            
                             event.editMessageEmbeds(getEmbed(this.page)).setActionRows(getButtonLayout(page)).queue();
                             waitForEvent(event.getChannel().getIdLong(), event.getMessageIdLong());
                             break;
@@ -167,14 +189,14 @@ public class EmbedPaginator {
                             if (page > pages) {
                                 page = pages;
                             }
-
+                            
                             event.editMessageEmbeds(getEmbed(this.page)).setActionRows(getButtonLayout(page)).queue();
                             waitForEvent(event.getChannel().getIdLong(), event.getMessageIdLong());
                             break;
 
                         case "stop":
                             interactionStopped = true;
-                            
+
                             if (event.getMessage() != null) {
                                 event.getMessage().delete().queue(s -> {}, e -> {});
                             } else {
@@ -201,7 +223,7 @@ public class EmbedPaginator {
                 () -> {
                     interactionStopped = true;
                     TextChannel channel = jda.getTextChannelById(channelId);
-                    
+
                     if (channel == null) {
                         return;
                     }
@@ -221,27 +243,49 @@ public class EmbedPaginator {
         if (page < 1) {
             page = 1;
         }
-        
-        int start = page == 1 ? 0 : ((page - 1));
-        int end = Math.min(items.size(), page);
-        
+
+        int start = page == 1 ? 0 : ((page - 1) * itemsPerPage);
+        int end = Math.min(items.size(), page * itemsPerPage);
+
+        StringBuilder itemsStringBuilder = new StringBuilder();
+
         for (int i = start; i < end; i++) {
-            this.items.get(i).setFooter("Page " + page + "/" + pages + (footer != null ? " • " + footer : ""));
+            itemsStringBuilder.append(numbered ? "`" + (i + 1) + ".` " : "").append(this.items.get(i)).append("\n");
         }
 
-        return this.items.get(page - 1).build();
+        EmbedBuilder builder = new EmbedBuilder()
+            .setTitle(embedTitle)
+            .setThumbnail(thumbnailUrl)
+            .setImage(imageUrl)
+            .setFooter("Page " + page + "/" + pages + (footer != null ? " • " + footer : ""))
+            .setColor(color)
+            .setDescription(embedDesc + "\r\n๑‧˚₊꒷꒦︶︶︶︶︶꒷꒦︶︶︶︶︶✦‧₊˚⊹\r\n" + itemsStringBuilder.toString().trim());
+
+        if (timestamp) {
+            builder.setTimestamp(Instant.now());
+        }
+        
+        return builder.build();
     }
 
     public static class Builder {
 
         private final JDA jda;
-        private final Set<Long> allowedUsers = new HashSet<>();
+        private final Set<Long> allowedUsers = new HashSet<Long>();
 
         private EventWaiter waiter;
         private long timeout = -1;
-        private List<EmbedBuilder> items;
+        private List<String> items;
+        private int itemsPerPage = 10;
+        private boolean numberItems = true;
         private String msgContent = null;
+        private Color color;
         private String footer;
+        private String embedTitle;
+        private String thumbnailUrl;
+        private String embedDesc;
+        private String imageUrl;
+        private boolean timestamp;
 
         public Builder(JDA jda) {
             this.jda = jda;
@@ -252,6 +296,31 @@ public class EmbedPaginator {
             return this;
         }
 
+        public Builder useTimestamp(boolean timestamp) {
+            this.timestamp = timestamp;
+            return this;
+        }
+
+        public Builder setImage(String imageUrl) {
+            this.imageUrl = imageUrl;
+            return this;
+        }
+
+        public Builder setEmbedDesc(String embedDesc) {
+            this.embedDesc = embedDesc;
+            return this;
+        }
+
+        public Builder setThumbnail(String thumbnailUrl) {
+            this.thumbnailUrl = thumbnailUrl;
+            return this;
+        }
+
+        public Builder setEmbedTitle(String embedTitle) {
+            this.embedTitle = embedTitle;
+            return this;
+        }
+
         public Builder setTimeout(long delay, TimeUnit unit) {
             Checks.notNull(unit, "TimeUnit");
             Checks.check(delay > 0, "Timeout must be greater than 0!");
@@ -259,7 +328,7 @@ public class EmbedPaginator {
             return this;
         }
 
-        public Builder setItems(List<EmbedBuilder> items) {
+        public Builder setItems(List<String> items) {
             this.items = items;
             return this;
         }
@@ -269,7 +338,28 @@ public class EmbedPaginator {
             return this;
         }
 
-        public Builder setMSgContent(String msgContent) {
+        public Builder setColor(Color color) {
+            this.color = color;
+            return this;
+        }
+
+        public Builder setColor(int color) {
+            this.color = EmbedUtils.intToColor(color);
+            return this;
+        }
+
+        public Builder setItemsPerPage(int items) {
+            Checks.check(items > 0, "Items per page must be at least 1");
+            this.itemsPerPage = items;
+            return this;
+        }
+
+        public Builder useNumberedItems(boolean b) {
+            this.numberItems = b;
+            return this;
+        }
+
+        public Builder setMsgContentContent(String msgContent) {
             this.msgContent = msgContent;
             return this;
         }
@@ -279,12 +369,14 @@ public class EmbedPaginator {
             return this;
         }
 
-        public EmbedPaginator build() {
+        public Paginator build() {
             Checks.notNull(waiter, "Waiter");
             Checks.check(timeout != -1, "You must set a timeout using #setTimeout()!");
-            Checks.noneNull(items, "items");
-            Checks.notEmpty(items, "items");
-            return new EmbedPaginator(waiter, timeout, items, jda, allowedUsers, msgContent, footer);
+            Checks.noneNull(items, "Items");
+            
+            return new Paginator(waiter, timeout, items, jda, allowedUsers, itemsPerPage,
+                                 numberItems, msgContent, color == null ? Color.black : color,
+                                 footer, embedTitle, thumbnailUrl, embedDesc, imageUrl, timestamp);
         }
     }       
 }
